+++ packages/theme/src/color-ramps/lib/constants.ts
@@ -33,8 +33,8 @@ export const ACCENT_SCALE_BASE_LIGHTNESS_THRESHOLDS = {
 	darker: { min: 0.75, max: 0.98 },
 } as const;
 
-// Lightness precision we aim for. Approximately 1/256, resolution of an 8-bit number.
-export const LIGHTNESS_EPSILON = 4e-3;
+// Contrast precision we aim for. Approximately 1/256, resolution of an 8-bit number.
+export const CONTRAST_EPSILON = 4e-3;
 
 export const MAX_BISECTION_ITERATIONS = 10;